+++ rt/lastmodifier/tests/test_viewlet.py
@@ -166,3 +166,11 @@ class TestViewlet(BaseTestCase):
         portal.folder.edit(description="Dolor Sit Amet")
         pq = PyQuery(portal.folder())
         self.assertTrue('by User 1' in pq('.documentModified').text())
+
+    def test_byline_on_site_root(self):
+        # Do not display anything
+        portal = self.layer['portal']
+        plone_utils = getToolByName(portal, 'plone_utils')
+        login(portal, TEST_USER_NAME)
+        pq = PyQuery(portal())
+        self.assertFalse('last modified' in pq('.documentModified').text())